% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_zip.R
\name{clean_zip}
\alias{clean_zip}
\title{Change zip codes with letters to NA values, and change those 3/4 digit zip codes to 5 digit form with prefix of "0"}
\usage{
clean_zip(x)
}
\arguments{
\item{x}{A character}
}
\value{
A character
}
\description{
Change zip codes with letters to NA values, and change those 3/4 digit zip codes to 5 digit form with prefix of "0"
}
