% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_zip.R
\name{contain_letter}
\alias{contain_letter}
\title{check if your zip code data includes letters due to data entry error.}
\usage{
contain_letter(x)
}
\arguments{
\item{x}{A character value to be checked if contain letters}
}
\value{
True/False
}
\description{
check if your zip code data includes letters due to data entry error.
}
