% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_zip.R
\name{get_zip}
\alias{get_zip}
\title{Extract zip codes from street adddress, for example "111 E Summit St, Kent, OH, 44240"}
\usage{
get_zip(x)
}
\arguments{
\item{x}{A character}
}
\value{
A character
}
\description{
Extract zip codes from street adddress, for example "111 E Summit St, Kent, OH, 44240"
}
